project/worker/requirements.txt
fastapi
uvicorn
psycopg2-binary
google-cloud-storage
python-dotenv
